% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_design_module.R
\name{analysisServer}
\alias{analysisServer}
\title{Experimental Design Analysis Server}
\usage{
analysisServer(id, home_inputs)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{home_inputs}{A reactive list from the home module.}
}
\description{
Contains all server logic for the EDA tabs.
}
