% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_design_module.R
\name{analysisUI}
\alias{analysisUI}
\title{Experimental Design Analysis UI}
\usage{
analysisUI(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\description{
This function defines the user interface for the two main analysis tabs ("Analysis 1" and "Analysis 2").
It returns a list of tabPanels to be inserted into the main app's navbarPage.
}
