% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mating_design_module.R
\name{create_explanation_ui}
\alias{create_explanation_ui}
\title{Create a Show/Hide UI for Explanations}
\usage{
create_explanation_ui(ns, base_id, content)
}
\arguments{
\item{ns}{The namespace function from the module server.}

\item{base_id}{A unique string to identify the UI elements.}

\item{content}{The HTML content to be shown or hidden.}
}
\value{
A tagList containing the action link and the hidden div.
}
\description{
A helper function to generate a consistent UI for showing and hiding
explanatory text content.
}
