% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_experiment_module.R
\name{custom_design_alpha}
\alias{custom_design_alpha}
\title{Generate an Alpha Lattice Design Layout}
\usage{
custom_design_alpha(treatments, k, r)
}
\arguments{
\item{treatments}{A character vector of treatment/genotype names.}

\item{k}{The number of plots per block (block size).}

\item{r}{The number of replications.}
}
\value{
A data frame representing the field book.
}
\description{
Creates an Alpha Lattice design field book from scratch.
}
