% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_experiment_module.R
\name{custom_design_rcbd}
\alias{custom_design_rcbd}
\title{Generate a Randomized Complete Block Design (RCBD) Layout}
\usage{
custom_design_rcbd(treatments, reps)
}
\arguments{
\item{treatments}{A character vector of treatment/genotype names.}

\item{reps}{The number of replications (blocks).}
}
\value{
A data frame representing the field book.
}
\description{
Creates an RCBD field book from a list of treatments and replications.
}
