% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_explorer_module.R
\name{dataCurationServer}
\alias{dataCurationServer}
\title{Data Curation Server (Sub-module)}
\usage{
dataCurationServer(id, home_inputs)
}
\arguments{
\item{id}{A unique identifier for the sub-module.}

\item{home_inputs}{A reactive list from the home module.}
}
\description{
Data Curation Server (Sub-module)
}
