% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_explorer_module.R
\name{dataCurationUI}
\alias{dataCurationUI}
\title{Data Curation UI (Sub-module)}
\usage{
dataCurationUI(id)
}
\arguments{
\item{id}{A unique identifier for the sub-module.}
}
\description{
Data Curation UI (Sub-module)
}
