% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_module.R
\name{homeServer}
\alias{homeServer}
\title{Home Tab Server}
\usage{
homeServer(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A reactive list containing the uploaded data and user selections.
}
\description{
This function defines the server-side logic for the Home tab. Its main
role is to listen for the user's request to proceed, gather all the inputs,
and return them as a reactive value to the main app.
}
