% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_module.R
\name{homeUI}
\alias{homeUI}
\title{Home Tab UI}
\usage{
homeUI(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\description{
This function defines the user interface for the main "Home" tab.
It includes all the input controls for selecting analysis types and uploading data.
}
