% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the PBAT Shiny app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}} (currently unused).}
}
\value{
A Shiny app object (in interactive sessions it will launch).
}
\description{
Sets up static resources bundled in the package and starts the app.
}
