% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pbat.R
\name{run_pbat}
\alias{run_pbat}
\title{Run the PBAT Shiny App}
\usage{
run_pbat(...)
}
\value{
Invisibly returns the app object.
}
\description{
Run the PBAT Shiny App
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
PBAT::run_pbat()
\dontshow{\}) # examplesIf}
}
