% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_explorer_module.R
\name{spatialExplorerServer}
\alias{spatialExplorerServer}
\title{Spatial Explorer Server (Sub-module)}
\usage{
spatialExplorerServer(id, home_inputs)
}
\arguments{
\item{id}{A unique identifier for the sub-module.}

\item{home_inputs}{A reactive list from the home module.}
}
\description{
Spatial Explorer Server (Sub-module)
}
