% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_explorer_module.R
\name{spatialExplorerUI}
\alias{spatialExplorerUI}
\title{Spatial Explorer UI (Sub-module)}
\usage{
spatialExplorerUI(id)
}
\arguments{
\item{id}{A unique identifier for the sub-module.}
}
\description{
Spatial Explorer UI (Sub-module)
}
