% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_explorer_module.R
\name{traitExplorerServer}
\alias{traitExplorerServer}
\title{Trait Explorer Main Server}
\usage{
traitExplorerServer(id, home_inputs)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{home_inputs}{A reactive list from the home module.}
}
\description{
This function calls the server logic for the two sub-modules.
}
