% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_explorer_module.R
\name{traitExplorerUI}
\alias{traitExplorerUI}
\title{Trait Explorer Main UI}
\usage{
traitExplorerUI(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\description{
This function creates a tabset panel to house the two sub-modules:
Spatial Trait Explorer and Data Curation.
}
